VoiceChatApp/requirements.txt
flask
flask_cors
speechrecognition
openai
python-dotenv
flask_socketio
requests